Protect pages base on permission

+++ routes/web.php
@@ -10,25 +10,25 @@ Route::inertia('/', 'welcome')->name('home');
 Route::middleware(['auth', 'verified'])->group(function () {
     Route::inertia('dashboard', 'dashboard')->name('dashboard');
 
-    Route::get('organization', [OrganizationController::class, 'index'])->name('organization');
-    Route::post('organization/post', [OrganizationController::class, 'post'])->name('proganization.post');
+    Route::get('organization', [OrganizationController::class, 'index'])->middleware('permission:organization.view')->name('organization');
+    Route::post('organization/post', [OrganizationController::class, 'post'])->middleware('permission:organization.view')->name('organization.post');
 
-    Route::prefix('roles')->name('roles')->group(function () {
+    Route::prefix('roles')->name('roles')->middleware('permission:roles.view')->group(function () {
         Route::get('/', [RoleController::class, 'index']);
-        Route::get('/create', [RoleController::class, 'create'])->name('.create');
-        Route::post('/', [RoleController::class, 'store'])->name('.store');
-        Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('.edit');
-        Route::put('/{role}', [RoleController::class, 'update'])->name('.update');
-        Route::delete('/{role}', [RoleController::class, 'delete'])->name('.delete');
+        Route::get('/create', [RoleController::class, 'create'])->name('.create')->middleware('permission:roles.create');
+        Route::post('/', [RoleController::class, 'store'])->name('.store')->middleware('permission:roles.create');
+        Route::get('/{role}/edit', [RoleController::class, 'edit'])->name('.edit')->middleware('permission:roles.edit');
+        Route::put('/{role}', [RoleController::class, 'update'])->name('.update')->middleware('permission:roles.edit');
+        Route::delete('/{role}', [RoleController::class, 'delete'])->name('.delete')->middleware('permission:roles.delete');
     });
 
-    Route::prefix('users')->name('users')->group(function () {
+    Route::prefix('users')->name('users')->middleware('permission:users.view')->group(function () {
         Route::get('/', [UserController::class, 'index']);
-        Route::get('/create', [UserController::class, 'create'])->name('.create');
-        Route::post('/', [UserController::class, 'store'])->name('.store');
-        Route::get('/{user}/edit', [UserController::class, 'edit'])->name('.edit');
-        Route::put('/{user}', [UserController::class, 'update'])->name('.update');
-        Route::delete('/{user}', [UserController::class, 'delete'])->name('.delete');
+        Route::get('/create', [UserController::class, 'create'])->name('.create')->middleware('permission:users.create');
+        Route::post('/', [UserController::class, 'store'])->name('.store')->middleware('permission:users.create');
+        Route::get('/{user}/edit', [UserController::class, 'edit'])->name('.edit')->middleware('permission:users.edit');
+        Route::put('/{user}', [UserController::class, 'update'])->name('.update')->middleware('permission:users.edit');
+        Route::delete('/{user}', [UserController::class, 'delete'])->name('.delete')->middleware('permission:users.delete');
     });
 
 });
